+++ src/main/java/io/github/astrapi69/gen/tree/handler/ITreeNodeHandlerExtensions.java
@@ -211,6 +211,27 @@ public class ITreeNodeHandlerExtensions
 	}
 
 	/**
+	 * Removes the given child from the given first {@link ITreeNode} Checks if the second given
+	 * {@link ITreeNode} object is a child of the first given {@link ITreeNode} object
+	 *
+	 * @param <V>
+	 *            the generic type of the value
+	 * @param <T>
+	 *            the generic type of the concrete tree node
+	 * @param parentTreeNode
+	 *            the parent tree node
+	 * @param child
+	 *            the child tree node to check
+	 * @return true, if the second given {@link ITreeNode} object is a child of the first
+	 *         {@link ITreeNode} object otherwise false
+	 */
+	public static <V, T extends ITreeNode<V, T>> boolean isChild(final @NonNull T parentTreeNode,
+		final T child)
+	{
+		return child != null && isChildOf(parentTreeNode, child);
+	}
+
+	/**
 	 * Checks if the second given {@link ITreeNode} object is a child of the first {@link ITreeNode}
 	 * object
 	 *